+++ __tests__/domain-whois-data/meabed-com-data.ts
@@ -1,0 +1,72 @@
+export const meabedComData = {
+  input: `
+     Domain Name: MEABED.COM
+   Registry Domain ID: 89119853_DOMAIN_COM-VRSN
+   Registrar WHOIS Server: whois.cloudflare.com
+   Registrar URL: http://www.cloudflare.com
+   Updated Date: 2021-11-24T21:17:37Z
+   Creation Date: 2002-08-07T12:32:41Z
+   Registry Expiry Date: 2026-08-07T12:32:41Z
+   Registrar: CloudFlare, Inc.
+   Registrar IANA ID: 1910
+   Registrar Abuse Contact Email:
+   Registrar Abuse Contact Phone:
+   Domain Status: clientTransferProhibited https://icann.org/epp#clientTransferProhibited
+   Name Server: LAKAS.NS.CLOUDFLARE.COM
+   Name Server: WALLY.NS.CLOUDFLARE.COM
+   DNSSEC: unsigned
+   URL of the ICANN Whois Inaccuracy Complaint Form: https://www.icann.org/wicf/
+>>> Last update of whois database: 2023-12-03T20:24:24Z <<<
+
+For more information on Whois status codes, please visit https://icann.org/epp
+
+NOTICE: The expiration date displayed in this record is the date the
+registrar's sponsorship of the domain name registration in the registry is
+currently set to expire. This date does not necessarily reflect the expiration
+date of the domain name registrant's agreement with the sponsoring
+registrar.  Users may consult the sponsoring registrar's Whois database to
+view the registrar's reported date of expiration for this registration.
+
+TERMS OF USE: You are not authorized to access or query our Whois
+database through the use of electronic processes that are high-volume and
+automated except as reasonably necessary to register domain names or
+modify existing registrations; the Data in VeriSign Global Registry
+Services' ("VeriSign") Whois database is provided by VeriSign for
+information purposes only, and to assist persons in obtaining information
+about or related to a domain name registration record. VeriSign does not
+guarantee its accuracy. By submitting a Whois query, you agree to abide
+by the following terms of use: You agree that you may use this Data only
+for lawful purposes and that under no circumstances will you use this Data
+to: (1) allow, enable, or otherwise support the transmission of mass
+unsolicited, commercial advertising or solicitations via e-mail, telephone,
+or facsimile; or (2) enable high volume, automated, electronic processes
+that apply to VeriSign (or its computer systems). The compilation,
+repackaging, dissemination or other use of this Data is expressly
+prohibited without the prior written consent of VeriSign. You agree not to
+use electronic processes that are automated and high-volume to access or
+query the Whois database except as reasonably necessary to register
+domain names or modify existing registrations. VeriSign reserves the right
+to restrict your access to the Whois database in its sole discretion to ensure
+operational stability.  VeriSign may restrict or terminate your access to the
+Whois database for failure to abide by these terms of use. VeriSign
+reserves the right to modify these terms at any time.
+
+The Registry database contains ONLY .COM, .NET, .EDU domains and
+Registrars.
+`,
+  output: {
+    domainName: 'meabed.com',
+    registrar: 'CloudFlare, Inc.',
+    updatedDate: '2021-01-24T21:17:37.000-05:00',
+    creationDate: '2002-01-07T12:32:41.000-05:00',
+    expirationDate: '2026-01-07T12:32:41.000-05:00',
+    status: ['clientTransferProhibited https://icann.org/epp#clientTransferProhibited'],
+    isAvailable: false,
+    nameServers: ['LAKAS.NS.CLOUDFLARE.COM', 'WALLY.NS.CLOUDFLARE.COM'],
+    registrarUrl: 'http://www.cloudflare.com',
+    registrarWhoisServer: 'whois.cloudflare.com',
+    registryDomainId: '89119853_DOMAIN_COM-VRSN',
+    dnssec: 'unsigned',
+    registrarIanaId: '1910',
+  },
+};